+++ src/sub/model_modeler_implementation/polynet/mpc_modeler_implementation_polynet.jl
@@ -22,7 +22,7 @@ Modeler implementation of a Model Predictive Control.
 """
 function _model_predictive_control_modeler_implementation(
     method::LinearProgramming,
-    model_mlj::AutomationLabsIdentification.PolyNet,
+    model_mlj::AutomationLabsSystems.PolyNet,
     system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem,
     horizon::Int,
     reference::ReferencesStateInput,
@@ -61,7 +61,7 @@ end
 
 function _model_predictive_control_modeler_implementation(
     method::NonLinearProgramming,
-    model_mlj::AutomationLabsIdentification.PolyNet,
+    model_mlj::AutomationLabsSystems.PolyNet,
     system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem,
     horizon::Int,
     reference::ReferencesStateInput,
@@ -200,7 +200,7 @@ end
 
 function _model_predictive_control_modeler_implementation(
     method::MixedIntegerLinearProgramming,
-    model_mlj::AutomationLabsIdentification.PolyNet,
+    model_mlj::AutomationLabsSystems.PolyNet,
     system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem,
     horizon::Int,
     reference::ReferencesStateInput,