+++ src/sub/model_modeler_implementation/resnet/mpc_modeler_implementation_resnet.jl
@@ -22,7 +22,7 @@ Modeler implementation of a Model Predictive Control.
 """
 function _model_predictive_control_modeler_implementation(
     method::LinearProgramming,
-    model_mlj::AutomationLabsIdentification.ResNet,
+    model_mlj::AutomationLabsSystems.ResNet,
     system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem,
     horizon::Int,
     reference::ReferencesStateInput,
@@ -61,7 +61,7 @@ end
 
 function _model_predictive_control_modeler_implementation(
     method::NonLinearProgramming,
-    model_mlj::AutomationLabsIdentification.ResNet,
+    model_mlj::AutomationLabsSystems.ResNet,
     system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem,
     horizon::Int,
     reference::ReferencesStateInput,
@@ -192,7 +192,7 @@ end
 
 function _model_predictive_control_modeler_implementation(
     method::MixedIntegerLinearProgramming,
-    model_mlj::AutomationLabsIdentification.ResNet,
+    model_mlj::AutomationLabsSystems.ResNet,
     system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem,
     horizon::Int,
     reference::ReferencesStateInput,
@@ -339,5 +339,4 @@ function _model_predictive_control_modeler_implementation(
     end
 
     return model_mpc
-
 end