+++ src/sub/model_modeler_implementation/rknn1/mpc_modeler_implementation_rknn1.jl
@@ -22,7 +22,7 @@ Modeler implementation of a Model Predictive Control.
 """
 function _model_predictive_control_modeler_implementation(
     method::LinearProgramming,
-    model_mlj::AutomationLabsIdentification.Rknn1,
+    model_mlj::AutomationLabsSystems.Rknn1,
     system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem,
     horizon::Int,
     reference::ReferencesStateInput,