+++ src/sub/model_modeler_implementation/rknn4/mpc_modeler_implementation_rknn4.jl
@@ -22,7 +22,7 @@ Modeler implementation of a Model Predictive Control.
 """
 function _model_predictive_control_modeler_implementation(
     method::LinearProgramming,
-    model_mlj::AutomationLabsIdentification.Rknn4,
+    model_mlj::AutomationLabsSystems.Rknn4,
     system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem,
     horizon::Int,
     reference::ReferencesStateInput,