+++ test/runtests.jl
@@ -23,13 +23,3 @@ println("done (took ", took_seconds, " seconds)")
 #print("Testing computation model predictive control...")
 #took_seconds = @elapsed include("./computation_mpc_test.jl");
 #println("done (took ", took_seconds, " seconds)")
-
-#=
-print("Testing design economic model predictive control...")
-took_seconds = @elapsed include("./design_empc_implementation_test.jl");
-println("done (took ", took_seconds, " seconds)")
-
-print("Testing computation economic model predictive control...")
-took_seconds = @elapsed include("./computation_empc_test.jl");
-println("done (took ", took_seconds, " seconds)")
-=#